+++ c#/imagePipeline/src/Ocr/RecognitionResults.cs
@@ -7,15 +7,15 @@ public interface IRecognitionResult
     public ImageKey ImageKey { get; }
     public RotatedRect TextBox { get; }
     public string Text { get; }
-    public ushort Confidence { get; }
+    public byte Confidence { get; }
 }
 
 public record PaddleOcrRecognitionResult
-    (ImageKey ImageKey, RotatedRect TextBox, string Text, ushort Confidence,
+    (ImageKey ImageKey, RotatedRect TextBox, string Text, byte Confidence,
         ModelVersion ModelVersion)
     : IRecognitionResult;
 
 public record TesseractRecognitionResult
-    (ImageKey ImageKey, RotatedRect TextBox, string Text, ushort Confidence,
+    (ImageKey ImageKey, RotatedRect TextBox, string Text, byte Confidence,
         bool IsVertical, bool ShouldFallbackToPaddleOcr)
     : IRecognitionResult;
